ref(ui): Consolidate server frontend hydration logic
